Part 4 finished

+++ part4/bloglist/controllers/blogs.js
@@ -1,40 +1,78 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(blogs.map(blog => blog.toJSON()))
 })
 
-blogsRouter.post('', async (request, response) => {
-  const blog = new Blog(request.body)
-
-  if (!blog.likes) {
-    blog.likes = 0
-  }
-
-  if (!blog.title && !blog.url) {
-    response.status(400).end()
-  } else {
-    try {
-      const savedBlog = await blog.save()
-      response.status(201).json(savedBlog.toJSON())
-    } catch (exception) {
-      console.log(exception)
-    }
-  }
-})
-
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.post('', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const token = request.token
+
+    if (!token) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = new Blog(request.body)
+
+    if (!blog.likes) {
+      blog.likes = 0
+    }
+
+    if (!blog.title && !blog.url) {
+      return response.status(400).json({ error: 'blog title and url missing' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    blog.user = user._id
+
+    const savedBlog = await blog.save()
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog.toJSON())
+
   } catch (exception) {
-    console.log(exception)
+    next(exception)
   }
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try {
+    const token = request.token
+    if (!token) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      return response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'unauthorized to delete blogs by other users' })
+    }
+  } catch (exception) {
+    next(exception)
+  }
+})
+
+blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
 
   if (!body.likes || (!body.title && !body.url)) {
@@ -55,7 +93,7 @@ blogsRouter.put('/:id', async (request, response) => {
         response.status(404).end()
       }
     } catch (exception) {
-      console.log(exception)
+      next(exception)
     }
 
   }
